@@ -10,24 +10,15 @@ import 'package:rubintv_visualization/workspace/data.dart';
 import 'package:rubintv_visualization/workspace/series.dart';
 import 'package:rubintv_visualization/workspace/window.dart';
 
-class UpdateBinsEvent extends ChartEvent {
-  final int nBins;
-
-  UpdateBinsEvent(this.nBins);
-}
-
 class BinnedState extends ChartStateLoaded {
   int nBins;
 
   BinnedState({
-    required super.key,
     required super.id,
     required super.series,
     required super.axisInfo,
     required super.legend,
     required super.useGlobalQuery,
-    required super.dataCenter,
-    required super.childKeys,
     required super.chartType,
     required super.tool,
     required this.nBins,
@@ -41,7 +32,6 @@ class BinnedState extends ChartStateLoaded {
     Legend? legend,
     bool? useGlobalQuery,
     DataCenter? dataCenter,
-    List<GlobalKey>? childKeys,
     WindowTypes? chartType,
     MultiSelectionTool? tool,
     int? nBins,
@@ -52,10 +42,7 @@ class BinnedState extends ChartStateLoaded {
         axisInfo: axisInfo ?? this.axisInfo,
         legend: legend ?? this.legend,
         useGlobalQuery: useGlobalQuery ?? this.useGlobalQuery,
-        dataCenter: dataCenter ?? this.dataCenter,
-        childKeys: childKeys ?? this.childKeys,
         chartType: chartType ?? this.chartType,
-        key: key,
         tool: tool ?? this.tool,
         nBins: nBins ?? this.nBins,
       );
@@ -72,7 +59,7 @@ class HistogramChart extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return BlocProvider(
-      create: (context) => ChartBloc(),
+      create: (context) => ChartBloc(window.id),
       child: BlocBuilder<ChartBloc, ChartState>(
         builder: (context, state) {
           if (state is! BinnedState) {
@@ -128,13 +115,11 @@ class HistogramChart extends StatelessWidget {
             ]),
             title: null,
             child: RubinChart(
-              key: state.key,
               info: HistogramInfo(
                 id: window.id,
                 allSeries: state.allSeries,
                 legend: state.legend,
                 axisInfo: state.axisInfo,
-                key: state.childKeys.first,
                 nBins: state.nBins,
               ),
               selectionController: selectionController,
