@@ -4,33 +4,64 @@ import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:rubin_chart/rubin_chart.dart';
 import 'package:rubintv_visualization/chart/base.dart';
+import 'package:rubintv_visualization/id.dart';
 import 'package:rubintv_visualization/state/workspace.dart';
 import 'package:rubintv_visualization/workspace/window.dart';
 
-class CartesianScatterPlot extends StatelessWidget {
+class InitializeScatterPlotEvent extends ChartEvent {
+  final UniqueId id;
+  final List<ChartAxisInfo> axisInfo;
+  final WindowTypes chartType;
+
+  InitializeScatterPlotEvent({
+    required this.id,
+    required this.axisInfo,
+    required this.chartType,
+  });
+}
+
+class ScatterPlotWidget extends StatelessWidget {
   final Window window;
 
-  const CartesianScatterPlot({
+  const ScatterPlotWidget({
     super.key,
     required this.window,
   });
 
   @override
   Widget build(BuildContext context) {
+    WorkspaceViewerState workspace = WorkspaceViewer.of(context);
+
     return BlocProvider(
-      create: (context) => ChartBloc(),
+      create: (context) {
+        List<ChartAxisInfo> axisInfo = [
+          ChartAxisInfo(
+            label: "<x>",
+            axisId: AxisId(AxisLocation.bottom),
+          ),
+          ChartAxisInfo(
+            label: "<y>",
+            axisId: AxisId(AxisLocation.left),
+            isInverted: true,
+          ),
+        ];
+        return ChartBloc(window.id)
+          ..add(InitializeScatterPlotEvent(
+            id: window.id,
+            axisInfo: axisInfo,
+            chartType: window.type,
+          ));
+      },
       child: BlocBuilder<ChartBloc, ChartState>(
         builder: (context, state) {
           if (state is! ChartStateLoaded) {
             /// Display an empty window while the chart is loading
             return ResizableWindow(
-                info: window,
-                title: "loading...",
-                toolbar:
-                    Row(children: [const Spacer(), ...context.read<ChartBloc>().getDefaultTools(context)]),
-                child: const Center(
-                  child: CircularProgressIndicator(),
-                ));
+              info: window,
+              title: "loading...",
+              toolbar: Row(children: [...context.read<ChartBloc>().getDefaultTools(context)]),
+              child: const Center(child: CircularProgressIndicator()),
+            );
           }
 
           WorkspaceViewerState workspace = WorkspaceViewer.of(context);
@@ -77,15 +108,21 @@ class CartesianScatterPlot extends StatelessWidget {
             ),
             title: null,
             child: RubinChart(
-              key: state.key,
-              info: CartesianScatterPlotInfo(
-                id: window.id,
-                allSeries: state.allSeries,
-                legend: state.legend,
-                axisInfo: state.axisInfo,
-                key: state.childKeys.first,
-                cursorAction: state.tool.cursorAction,
-              ),
+              info: window.type == WindowTypes.cartesianScatter
+                  ? CartesianScatterPlotInfo(
+                      id: window.id,
+                      allSeries: state.allSeries,
+                      legend: state.legend,
+                      axisInfo: state.axisInfo,
+                      cursorAction: state.tool.cursorAction,
+                    )
+                  : PolarScatterPlotInfo(
+                      id: window.id,
+                      allSeries: state.allSeries,
+                      legend: state.legend,
+                      axisInfo: state.axisInfo,
+                      cursorAction: state.tool.cursorAction,
+                    ),
               selectionController: selectionController,
               drillDownController: drillDownController,
             ),
